+++ app_wsbrd/ws/ws_common.h
@@ -72,6 +72,7 @@ typedef struct ws_info {
     struct ws_phy_config phy_config;
     struct ws_neigh_table neighbor_storage;
     struct ws_fhss_config fhss_config;
+    struct timer_entry bc_resync_timer;
     uint32_t tx_duration_ms; // Time spent in TX during the last hour
     uint8_t ffn_gtk_index;
     uint8_t lfn_gtk_index;